+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added the initial basis to support different ports on different targets
 - Now is possible to specify the port on targets list (ex: 10.10.1.10:2222) (see #5)
 - Shows time elapsed on each phase.
+- Remote command execution (argument: -X <command>).
 - Increase the maximum file descriptor number that can be opened by this process.
 - manpages (`man cbrutekrag`)
 - Debug bracktrace symbols